Updates for imports api validation
